Update class location. Moved packages
