+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/execution/ScriptExecutionFilterInput.java
@@ -1,4 +1,4 @@
-package com.openframe.api.dto.execution;
+package com.openframe.api.dto.rmm.execution;
 
 import com.openframe.data.document.rmm.ExecutionStatus;
 import lombok.AllArgsConstructor;
