+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/execution/ScriptExecutionFilters.java
@@ -1,7 +1,6 @@
-package com.openframe.api.dto.execution;
+package com.openframe.api.dto.rmm.execution;
 
-import com.openframe.api.dto.script.ScriptFilterOption;
-import com.openframe.data.document.rmm.ExecutionStatus;
+import com.openframe.api.dto.rmm.script.ScriptFilterOption;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
