+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/script/RunScriptInput.java
@@ -1,4 +1,4 @@
-package com.openframe.api.dto.script;
+package com.openframe.api.dto.rmm.script;
 
 import com.openframe.data.document.rmm.PrivilegeLevel;
 import jakarta.validation.Valid;
