+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/script/ScriptEnvVarInput.java
@@ -1,4 +1,4 @@
-package com.openframe.api.dto.script;
+package com.openframe.api.dto.rmm.script;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
