+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/script/ScriptFilterOption.java
@@ -1,4 +1,4 @@
-package com.openframe.api.dto.script;
+package com.openframe.api.dto.rmm.script;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
