+++ openframe-api-lib/src/main/java/com/openframe/api/dto/rmm/script/ScriptResponse.java
@@ -1,4 +1,4 @@
-package com.openframe.api.dto.script;
+package com.openframe.api.dto.rmm.script;
 
 import com.openframe.data.document.rmm.PrivilegeLevel;
 import lombok.Builder;
