+++ openframe-api-service-core/src/main/java/com/openframe/api/dataloader/ScriptDataLoader.java
@@ -1,7 +1,7 @@
 package com.openframe.api.dataloader;
 
 import com.netflix.graphql.dgs.DgsDataLoader;
-import com.openframe.api.dto.script.ScriptResponse;
+import com.openframe.api.dto.rmm.script.ScriptResponse;
 import com.openframe.api.service.rmm.ScriptService;
 import lombok.RequiredArgsConstructor;
 import org.dataloader.BatchLoader;
