+++ openframe-api-service-core/src/main/java/com/openframe/api/mapper/GraphQLScriptExecutionMapper.java
@@ -3,7 +3,7 @@ package com.openframe.api.mapper;
 import com.openframe.api.dto.CountedGenericConnection;
 import com.openframe.api.dto.CountedGenericQueryResult;
 import com.openframe.api.dto.GenericEdge;
-import com.openframe.api.dto.execution.ScriptExecutionResponse;
+import com.openframe.api.dto.rmm.execution.ScriptExecutionResponse;
 import com.openframe.api.dto.shared.ConnectionArgs;
 import com.openframe.api.dto.shared.CursorCodec;
 import com.openframe.api.dto.shared.CursorPaginationCriteria;
