+++ openframe-api-service-core/src/main/java/com/openframe/api/relay/NodeTypeResolver.java
@@ -2,8 +2,8 @@ package com.openframe.api.relay;
 
 import com.netflix.graphql.dgs.DgsComponent;
 import com.netflix.graphql.dgs.DgsTypeResolver;
-import com.openframe.api.dto.execution.ScriptExecutionResponse;
-import com.openframe.api.dto.script.ScriptResponse;
+import com.openframe.api.dto.rmm.execution.ScriptExecutionResponse;
+import com.openframe.api.dto.rmm.script.ScriptResponse;
 import com.openframe.data.document.assignment.ItemAssignment;
 import com.openframe.data.document.device.Machine;
 import com.openframe.data.document.knowledgebase.KnowledgeBaseItem;
